requirements.txt
python-dotenv==1.0.0
pytrends==4.9.2
openai==1.3.7
tweepy==4.14.0
reportlab==4.0.7
sendgrid==6.10.0
flask==3.0.0
flask-cors==4.0.0
